refactor: support virtual bb.environment label without explicit values in `labels` object
