+++ frontend/src/utils/v1/index.ts
@@ -4,11 +4,9 @@ export * from "./instance";
 export * from "./database";
 export * from "./databaseGroup";
 export * from "./backup";
-export * from "./databaseLabel";
 export * from "./project";
 export * from "./projectWebhook";
 export * from "./deploymentConfig";
-export * from "./label";
 export * from "./role";
 export * from "./user";
 export * from "./instance";
